\name{globtp}
\alias{globtp}
\docType{data}
\title{Annual global temperature, 1880-1985}

\description{
Surface air temperature departure from from a set value.
}
\usage{data(globtp)}
\format{
  The format is:
 Time-Series [1:106] from 1880 to 1985: -0.4 -0.37 -0.43 -0.47 -0.72 -0.54 -0.47 -0.54 -0.39 -0.19 ...
}

\source{
James Hansen and Sergej Lebedeff, Global Trends of Measured Surface Air Temperature, Journal of Geophysical Research, Vol. 92, No. D11, pages 13,345-13,372, November 20, 1987.
}

\examples{
data(globtp)
plot(globtp)
}
\keyword{datasets}
